channel support added

+++ db_postgresql/init.sql
@@ -1,6 +1,6 @@
 CREATE TABLE conversations (
     id SERIAL PRIMARY KEY,
     conv_text varchar(1000),
-    chat_id INTEGER,
+    chat_id bigint,
     updated_at TIMESTAMP
 );
